DEV: extract updating status on mentions into a lib function
