+++ app/assets/javascripts/discourse/app/widgets/post-cooked.js
@@ -4,13 +4,12 @@ import { ajax } from "discourse/lib/ajax";
 import highlightSearch from "discourse/lib/highlight-search";
 import { iconHTML } from "discourse-common/lib/icon-library";
 import { isValidLink } from "discourse/lib/click-track";
-import { number, until } from "discourse/lib/formatter";
+import { number } from "discourse/lib/formatter";
 import { spinnerHTML } from "discourse/helpers/loading-spinner";
 import { escape } from "pretty-text/sanitizer";
 import domFromString from "discourse-common/lib/dom-from-string";
 import getURL from "discourse-common/lib/get-url";
-import { emojiUnescape } from "discourse/lib/text";
-import { escapeExpression } from "discourse/lib/utilities";
+import { updateUserStatusOnMention } from "discourse/lib/update-user-status-on-mention";
 
 let _beforeAdoptDecorators = [];
 let _afterAdoptDecorators = [];
@@ -392,41 +391,8 @@ export default class PostCooked {
     const mentions = postElement.querySelectorAll(`a.mention[href="${href}"]`);
 
     mentions.forEach((mention) => {
-      this._updateUserStatus(mention, user.status);
+      updateUserStatusOnMention(mention, user.status, this.currentUser);
     });
-  }
-
-  _updateUserStatus(mention, status) {
-    this._removeUserStatus(mention);
-    if (status) {
-      this._insertUserStatus(mention, status);
-    }
-  }
-
-  _insertUserStatus(mention, status) {
-    const emoji = escapeExpression(`:${status.emoji}:`);
-    const statusHtml = emojiUnescape(emoji, {
-      class: "user-status",
-      title: this._userStatusTitle(status),
-    });
-    mention.insertAdjacentHTML("beforeend", statusHtml);
-  }
-
-  _removeUserStatus(mention) {
-    mention.querySelector("img.user-status")?.remove();
-  }
-
-  _userStatusTitle(status) {
-    if (!status.ends_at) {
-      return status.description;
-    }
-
-    const timezone = this.currentUser
-      ? this.currentUser.user_option?.timezone
-      : moment.tz.guess();
-
-    const until_ = until(status.ends_at, timezone, this.currentUser?.locale);
-    return escapeExpression(`${status.description} ${until_}`);
   }
 
   _trackMentionedUsersStatus() {
